+++ src/integrations/elements/Entry.php
@@ -401,6 +401,7 @@ class Entry extends Element
             'label' => Craft::t('formie', 'Entry Type'),
             'instructions' => Craft::t('formie', 'Select an entry type to map content to. This will reflect the available fields to map to.'),
             'required' => true,
+            'placeholder' => Craft::t('formie', 'Select an option'),
             'options' => $this->_getEntryTypeOptions(),
         ]);
         $schema[] = SchemaHelper::lightswitchField([
@@ -508,7 +509,7 @@ class Entry extends Element
 
     private function _getEntryTypeOptions(): array
     {
-        $options = [['label' => Craft::t('formie', 'Select an option'), 'value' => '']];
+        $options = [];
         $elements = $this->getFormSettingValue('elements');
 
         if (!is_array($elements)) {