+++ src/integrations/elements/Product.php
@@ -362,6 +362,7 @@ class Product extends Element
             'label' => Craft::t('formie', 'Product Type'),
             'instructions' => Craft::t('formie', 'Select a product type to map content to. This will reflect the available fields to map to.'),
             'required' => true,
+            'placeholder' => Craft::t('formie', 'Select an option'),
             'options' => $this->_getProductTypeOptions(),
         ]);
         $schema[] = SchemaHelper::elementSelectField([
@@ -442,7 +443,7 @@ class Product extends Element
 
     private function _getProductTypeOptions(): array
     {
-        $options = [['label' => Craft::t('formie', 'Select an option'), 'value' => '']];
+        $options = [];
         $elements = $this->getFormSettingValue('elements');
         if (is_array($elements)) {
             foreach ($elements as $item) {